// blazor-grid-components/scripts/interfaces.ts
export interface ISelectedCell {
    rowIndex: number;
    cellIndexes: number[];
}